feat(upsell): render upgrade content in-tab for AI Shell & ER Diagram
Community users opening an AI Shell or ER Diagram plugin tab now see the same upgrade content the modal shows — but rendered inline as the tab body, with the matching feature pre-selected. The standalone AiShellUpsell card is gone in favour of this shared content.

- New UpgradeContent.vue holds the entire modal body (header, feature picker, hero, footer, lifetime band). Accepts trigger-feature / custom-title / message / in-tab / closable props. Internal active state, watches trigger-feature so a reopen / re-trigger jumps to the right tab.
- New upgrade-features.ts factors the FEATURES array + inferFeature() out of UpgradeRequiredModal — both consumers share the same source of truth.
- UpgradeRequiredModal.vue is now a thin shell: wraps <UpgradeContent closable> inside the vue-js-modal, plus the show/hide event plumbing and the global keyboard-shortcut trap.
- TabPluginShell.vue renders <UpgradeContent trigger-feature="ai" in-tab> for community AI tabs and <UpgradeContent trigger-feature="erd" in-tab> for ERD tabs, dropping the previous AppEvent.upgradeModal hack that opened the modal on top of the tab.
- AiShellUpsell.vue and components/upsell/ai-shell-upsell.scss are deleted; the AiShellPreview animation is still used as the AI tab's preview inside UpgradeContent.
- modals.scss: chrome (.v--modal sizing) stays under .beekeeper-modal.upgrade-modal; every .um-* / .ump-* rule moves to a top-level .upgrade-content block so the same styles apply inside both the modal and the inline tab. New .upgrade-content.in-tab variant fills the wrapper.
- core-tabs.scss: rename .tab-upsell-wrapper--ai-shell to --full (used for both AI Shell and ER Diagram now).
