+++ apps/studio/src/assets/styles/app/core-tabs.scss
@@ -566,7 +566,8 @@
 .tab-upsell-wrapper {
   padding: 0rem 2rem 2rem;
 
-  &.tab-upsell-wrapper--ai-shell {
+  // Full-bleed mode for in-tab UpgradeContent (AI Shell, ER Diagram).
+  &.tab-upsell-wrapper--full {
     padding: 0;
     height: 100%;
     min-height: 0;
