+++ apps/studio/src/assets/styles/components/_all.scss
@@ -8,4 +8,3 @@
 @import "text-editor.scss";
 @import "json-viewer.scss";
 @import './upsell/upsell-content.scss';
-@import './upsell/ai-shell-upsell.scss';
